PS2-149 - Karna ArticleVat is a long instead of Decimal
Karna ArticleVat is a long instead of Decimal

+++ BuckarooSdk/DataTypes/ParameterGroups/Klarna/Article.cs
@@ -2,14 +2,14 @@ using BuckarooSdk.Services;
 
 namespace BuckarooSdk.DataTypes.ParameterGroups.Klarna
 {
-	public class Article : ParameterGroup
+    public class Article : ParameterGroup
 	{
 		public string ArticleTitle { get; set; }
 		public string ArticleNumber { get; set; }
 		public string ArticleType { get; set; }
 		public int ArticleQuantity { get; set; }
 		public decimal ArticlePrice { get; set; }
-		public long ArticleVat { get; set; }
+		public decimal ArticleVat { get; set; }
 		public decimal ArticleDiscount { get; set; }
 	}
 }
